0-linear.c hot fix

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,16 +1,17 @@
 #include "search_algos.h"
 
 /**
- * linear_search - searches for a value in an array of integers
- * @array: array of integers
- * @size: size of array
- * @value: value that should be found
+ * linear_search - searches for a value in an array of
+ * integers using the Linear search algorithm
  *
- * Return: index of value searched or -1 if it doesn't exist
+ * @array: input array
+ * @size: size of the array
+ * @value: value to search in
+ * Return: Always EXIT_SUCCESS
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (array == NULL)
 		return (-1);
@@ -21,6 +22,5 @@ int linear_search(int *array, size_t size, int value)
 		if (value == array[i])
 			return (i);
 	}
-
 	return (-1);
 }
